move packages from cmd/* to cmd/funcs/*
