+++ pkg/cmd/funcs/authstatus/read.go
