+++ pkg/cmd/funcs/authstatus/read_gen.go
