+++ pkg/cmd/funcs/authstatus/resource.go
