+++ pkg/cmd/funcs/disk/create.go
