+++ pkg/cmd/funcs/disk/create_test.go
