+++ pkg/cmd/funcs/disk/list.go
