+++ pkg/cmd/funcs/disk/list_gen.go
