+++ pkg/cmd/funcs/disk/list_test.go
