+++ pkg/cmd/funcs/disk/resource.go
