+++ pkg/cmd/funcs/disk/update.go
