+++ pkg/cmd/funcs/disk/update_test.go
